global total added to Bill edit screen

+++ db/migrate/20110201071016_add_total_to_products.rb
@@ -1,7 +1,7 @@
 class AddTotalToProducts < ActiveRecord::Migration
   def self.up
-    add_column :products, :totaltvac, :integer
-    add_column :products, :totalhtva, :integer
+    add_column :products, :totaltvac, :float
+    add_column :products, :totalhtva, :float
   end
 
   def self.down
